DACN/backend_src/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
python-jose[cryptography]
passlib[bcrypt]
pydantic
pydantic-settings
opencv-python
psycopg2-binary
numpy
tensorflow
mysql-connector-python
python-multipart
requests
loguru
python-dotenv
pytest
pytest-asyncio
httpx
slowapi
flask
Pillow
redis
pytest-cov
bleach
cryptography
